requirements.txt
python-telegram-bot==13.15
APScheduler==3.9.1
python-dotenv==0.21.0
requests==2.31.0